+++ src/net/dxzc/pome/PointBuffer.java
@@ -4,23 +4,23 @@ public final class PointBuffer {
 
     public Mesh mesh;
 
-    public float buffer1;
+    public double buffer1;
 
-    public float buffer2;
+    public double buffer2;
 
-    public float buffer3;
+    public double buffer3;
 
-    public float buffer4;
+    public double buffer4;
 
-    public float buffer5;
+    public double buffer5;
 
-    public float buffer6;
+    public double buffer6;
 
-    public float buffer7;
+    public double buffer7;
 
-    public float buffer8;
+    public double buffer8;
 
-    public float buffer9;
+    public double buffer9;
 
     public void clear() {
         mesh = null;
