+++ src/net/dxzc/pome/PointData.java
@@ -2,41 +2,41 @@ package net.dxzc.pome;
 
 public final class PointData {
 
-    public float x;
+    public double x;
 
-    public float y;
+    public double y;
 
-    public float z;
+    public double z;
 
-    public float normalX;
+    public double normalX;
 
-    public float normalY;
+    public double normalY;
 
-    public float normalZ;
+    public double normalZ;
 
-    public float tangentX;
+    public double tangentX;
 
-    public float tangentY;
+    public double tangentY;
 
-    public float tangentZ;
+    public double tangentZ;
 
-    public float buffer1;
+    public double buffer1;
 
-    public float buffer2;
+    public double buffer2;
 
-    public float buffer3;
+    public double buffer3;
 
-    public float buffer4;
+    public double buffer4;
 
-    public float buffer5;
+    public double buffer5;
 
-    public float buffer6;
+    public double buffer6;
 
-    public float buffer7;
+    public double buffer7;
 
-    public float buffer8;
+    public double buffer8;
 
-    public float buffer9;
+    public double buffer9;
 
     public void set(PointData a) {
         x = a.x;
@@ -59,7 +59,7 @@ public final class PointData {
         buffer9 = a.buffer9;
     }
 
-    public void set(PointData a, float ap) {
+    public void set(PointData a, double ap) {
         buffer1 = a.buffer1 * ap;
         buffer2 = a.buffer2 * ap;
         buffer3 = a.buffer4 * ap;
@@ -71,7 +71,7 @@ public final class PointData {
         buffer9 = a.buffer9 * ap;
     }
 
-    public void set(PointData a, PointData b, float ap, float bp) {
+    public void set(PointData a, PointData b, double ap, double bp) {
         buffer1 = a.buffer1 * ap + b.buffer1 * bp;
         buffer2 = a.buffer2 * ap + b.buffer2 * bp;
         buffer3 = a.buffer4 * ap + b.buffer3 * bp;
@@ -83,7 +83,7 @@ public final class PointData {
         buffer9 = a.buffer9 * ap + b.buffer9 * bp;
     }
 
-    public void set(PointData a, PointData b, PointData c, float ap, float bp, float cp) {
+    public void set(PointData a, PointData b, PointData c, double ap, double bp, double cp) {
         buffer1 = a.buffer1 * ap + b.buffer1 * bp + c.buffer1 * cp;
         buffer2 = a.buffer2 * ap + b.buffer2 * bp + c.buffer2 * cp;
         buffer3 = a.buffer4 * ap + b.buffer3 * bp + c.buffer3 * cp;
